新增邮箱模块

+++ clever-system/src/main/java/com/clever/service/impl/EmailSubjectServiceImpl.java
@@ -2,13 +2,22 @@ package com.clever.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
+import com.clever.Constant;
 import com.clever.bean.model.OnlineUser;
+import com.clever.bean.system.EmailTemplate;
+import com.clever.exception.BaseException;
+import com.clever.exception.ConstantException;
+import com.clever.manager.EmailManager;
+import com.clever.manager.JavaMailSenderFactory;
+import com.clever.mapper.EmailTemplateMapper;
+import com.clever.util.DesUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 import com.clever.mapper.EmailSubjectMapper;
 import com.clever.bean.system.EmailSubject;
@@ -29,6 +38,10 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
 
     @Resource
     private EmailSubjectMapper emailSubjectMapper;
+
+    @Resource
+    private EmailTemplateMapper emailTemplateMapper;
+
 
     /**
      * 分页查询邮箱主体列表
@@ -101,6 +114,10 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public EmailSubject create(EmailSubject emailSubject, OnlineUser onlineUser) {
+        if (emailSubjectMapper.selectCount(getEmailSubjectQueryWrapper(emailSubject, onlineUser)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("您已存在同一个邮箱主体"));
+        }
+        emailSubject.setPassword(DesUtil.safeEncrypt(emailSubject.getPassword(), Constant.KEY));
         emailSubjectMapper.insert(emailSubject);
         log.info("邮箱主体, 邮箱主体信息创建成功: userId={}, emailSubjectId={}", onlineUser.getId(), emailSubject.getId());
         return emailSubject;
@@ -115,7 +132,21 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public EmailSubject update(EmailSubject emailSubject, OnlineUser onlineUser) {
+        QueryWrapper<EmailSubject> queryWrapper = getEmailSubjectQueryWrapper(emailSubject, onlineUser);
+        queryWrapper.ne("id", emailSubject.getId());
+        if (emailSubjectMapper.selectCount(getEmailSubjectQueryWrapper(emailSubject, onlineUser)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("您已存在同一个邮箱主体"));
+        }
+        EmailSubject oldEmailSubject = emailSubjectMapper.selectById(emailSubject.getId());
+        if (oldEmailSubject == null) {
+            throw new BaseException(ConstantException.DATA_NOT_EXIST.format("邮箱主体"));
+        }
+        // 防止重复加密
+        if (!oldEmailSubject.getPassword().equals(emailSubject.getPassword())) {
+            emailSubject.setPassword(DesUtil.safeEncrypt(emailSubject.getPassword(), Constant.KEY));
+        }
         emailSubjectMapper.updateById(emailSubject);
+        JavaMailSenderFactory.updateMailSender(emailSubject.getId()); // 更新邮箱主体
         log.info("邮箱主体, 邮箱主体信息修改成功: userId={}, emailSubjectId={}", onlineUser.getId(), emailSubject.getId());
         return emailSubject;
     }
@@ -143,6 +174,9 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void delete(String id, OnlineUser onlineUser) {
+        if (emailTemplateMapper.selectCount(new QueryWrapper<EmailTemplate>().eq("subject_id", id)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("该邮箱主体下还有邮件模板正在使用!!!"));
+        }
         emailSubjectMapper.deleteById(id);
         log.info("邮箱主体, 邮箱主体信息删除成功: userId={}, emailSubjectId={}", onlineUser.getId(), id);
     }
@@ -155,6 +189,9 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void deleteBatchIds(List<String> ids, OnlineUser onlineUser) {
+        if (emailTemplateMapper.selectCount(new QueryWrapper<EmailTemplate>().in("subject_id", ids)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("存在邮箱主体下还有邮件模板正在使用!!!"));
+        }
         emailSubjectMapper.deleteBatchIds(ids);
         log.info("邮箱主体, 邮箱主体信息批量删除成功: userId={}, count={}, emailSubjectIds={}", onlineUser.getId(), ids.size(), ids.toString());
     }
@@ -167,7 +204,12 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void deleteByPlatformId(Integer platformId, OnlineUser onlineUser) {
-        emailSubjectMapper.delete(new QueryWrapper<EmailSubject>().eq("platform_id", platformId));
+        List<EmailSubject> emailSubjects = emailSubjectMapper.selectList(new QueryWrapper<EmailSubject>().eq("platform_id", platformId));
+        List<String> ids = emailSubjects.stream().map(EmailSubject::getId).collect(Collectors.toList());
+        if (emailTemplateMapper.selectCount(new QueryWrapper<EmailTemplate>().in("subject_id", ids)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("存在邮箱主体下还有邮件模板正在使用!!!"));
+        }
+        emailSubjectMapper.deleteBatchIds(ids);
         log.info("邮箱主体, 邮箱主体信息根据platformId删除成功: userId={}, platformId={}", onlineUser.getId(), platformId);
     }
 
@@ -179,7 +221,25 @@ public class EmailSubjectServiceImpl implements EmailSubjectService {
      */
     @Override
     public void deleteByCreator(String creator, OnlineUser onlineUser) {
-        emailSubjectMapper.delete(new QueryWrapper<EmailSubject>().eq("creator", creator));
+        List<EmailSubject> emailSubjects = emailSubjectMapper.selectList(new QueryWrapper<EmailSubject>().eq("creator", creator));
+        List<String> ids = emailSubjects.stream().map(EmailSubject::getId).collect(Collectors.toList());
+        if (emailTemplateMapper.selectCount(new QueryWrapper<EmailTemplate>().in("subject_id", ids)) > 0) {
+            throw new BaseException(ConstantException.DATA_IS_EXIST.reset("存在邮箱主体下还有邮件模板正在使用!!!"));
+        }
+        emailSubjectMapper.deleteBatchIds(ids);
         log.info("邮箱主体, 邮箱主体信息根据creator删除成功: userId={}, creator={}", onlineUser.getId(), creator);
+    }
+
+    private static QueryWrapper<EmailSubject> getEmailSubjectQueryWrapper(EmailSubject emailSubject, OnlineUser onlineUser) {
+        QueryWrapper<EmailSubject> queryWrapper = new QueryWrapper<>();
+        queryWrapper.eq("platform_id", emailSubject.getPlatformId());
+        queryWrapper.eq("creator", onlineUser.getId());
+        queryWrapper.eq("host", emailSubject.getHost());
+        queryWrapper.eq("driver", emailSubject.getDriver());
+        queryWrapper.eq("port", emailSubject.getPort());
+        queryWrapper.eq("encryption", emailSubject.getEncryption());
+        queryWrapper.eq("username", emailSubject.getUsername());
+        queryWrapper.eq("subject_name", emailSubject.getSubjectName());
+        return queryWrapper;
     }
 }
